requirements.txt
git+https://github.com/JustAnotherArchivist/snscrape.git
google-cloud-videointelligence
videocr